ANDROID: overlayfs: add __get xattr method
Because of the overlayfs getxattr recursion, the incoming inode fails
to update the selinux sid resulting in avc denials being reported
against a target context of u:object_r:unlabeled:s0.

Solution is to add a _get xattr method that calls the __vfs_getxattr
handler so that the context can be read in, rather than being denied
with an -EACCES when vfs_getxattr handler is called.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 133515582
Bug: 136124883
Bug: 129319403
Change-Id: Ia39543c5ce617976f14d790fb88e471d575ffd65
Git-commit: 18cb05fae20ff6df825141414e607d06c30c99cc
Git-repo: https://android.googlesource.com/kernel/common
[jshriram@codeaurora.org: No changes]
Signed-off-by: Jeevan Shriram <jshriram@codeaurora.org>
[groverm@codeaurora.org: No changes]
Signed-off-by: Mayank Grover <groverm@codeaurora.org>