[BACKPORT] ext4/f2fs: migrate to generic crypto facility
Actually we are using an old (specific for ext4) crypto facility
This is a problem, for at least 3 reasons:

1 - Kernel sometimes crashes due to a NULL pointer deference, I wasn't able
to debug that (happened when I had debugging disabled), but after I disassembled
the kernel I noticed that the crash happens somewhere in crypto facility
2 - We do not support FBE V1, only ICE encryption
3 - This prevents f2fs to work, because cryptfs lacks a ton of updates that cannot
be merged, changes will break the crypto facility we are actually using on ext4.
Additionally FBE support is only partially implemented in f2fs driver, and it's impossible
to add it in this situation (unless someone wants to create it from scratch on a dead kernel)

On 4.9 kernel both ext4 and f2fs drivers use the generic crypto facility and works pretty fine
as none of my users reported a single issue on ext4 and f2fs. That's why i decided to port
all the needed code from 4.9 kernel.
Unfortunately it was impossible to keep track of all changes during the porting, here's a
partial list of the most important things merged and adapted for 4.4 kernel:

PFE:
- sync security/pfe with 4.9 kernel
- crypto: Remove crypto block device from encryption path
- ice: Add code to support older implementation
- ice: msm: Encrypt only userdata
- security: pfe: make inline crypto engine scm call based on version

EXT4:
- ext4 crypto: migrate into vfs's crypto engine
- ext4: adapt code to fscrypt changes
- ext4: switch to fscrypt ->symlink() helper functions
- Merged some missing stuffs from fscrypt v4.17-rc1

F2FS
- fscrypt: remove filesystem specific build config option
- Enable hardware based FBE on f2fs and adapt ext4 fs

FSCRYPT:
- fscrypt: remove filesystem specific build config option
- fscrypt: remove 'ci' parameter from fscrypt_put_encryption_info()
- fscrypt: allow synchronous bio decryption
- Update Inline Crypto Engine support for File based Encryption
- fscrypto: make filename crypto functions return 0 on success
- fscrypt: updates to v4.17-rc1
- fscrypt: make ->dummy_context() return bool

MISC:
- Resolve merge conflict and enable HW FBE for ext4 fs <-- this is a rebase of the commit
we have in 4.4 kernel made over 4.9 kernel updates
- Enable hardware based FBE on f2fs and adapt ext4
- Fixed several spacing/identation issues
- Restored some code that has been backported only partially on 4.4 kernel (it was probably
breaking something in old crypto facility)

TEST: added FBE V1 support in device tree, build, boot, tested all functions with both
EXT4 and F2FS

[huexxx@gmail.com: adapt]
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>